Resolved CI issues and few UI bugs

+++ app/server/controllers/nfs.js
@@ -1,6 +1,6 @@
 import mime from 'mime';
 import sessionManager from '../session_manager';
-import { formatResponse, ResponseError, ResponseHandler } from '../utils';
+import { formatDirectoryResponse, formatResponse, ResponseError, ResponseHandler } from '../utils';
 import { log } from './../../logger/log';
 import { NfsWriter } from '../stream/nfs_writer';
 import { NfsReader } from '../stream/nfs_reader';
@@ -29,16 +29,25 @@ let deleteOrGetDirectory = function(req, res, isDelete, next) {
   if (typeof rootPath === 'undefined') {
     return next(new ResponseError(400, util.format(MSG_CONSTANTS.FAILURE.FIELD_NOT_VALID, 'rootPath')));
   }
-  let dirPath = req.params['0'];
+  let dirPath = req.params['0'] || '/';
   let responseHandler = new ResponseHandler(req, res);
   if (isDelete) {
+    if (dirPath === '/') {
+      return next(new ResponseError(400, 'Cannot delete root directory'))
+    }
     log.debug('NFS - Invoking delete directory request');
     req.app.get('api').nfs.deleteDirectory(dirPath, rootPath,
       sessionInfo.hasSafeDriveAccess(), sessionInfo.appDirKey, responseHandler);
   } else {
     log.debug('NFS  - Invoking get directory request');
     req.app.get('api').nfs.getDirectory(dirPath, rootPath,
-      sessionInfo.hasSafeDriveAccess(), sessionInfo.appDirKey, responseHandler);
+      sessionInfo.hasSafeDriveAccess(), sessionInfo.appDirKey, function(err, dir) {
+        if (err) {
+          return responseHandler(err);
+        }
+        dir = formatDirectoryResponse(JSON.parse(dir));
+        responseHandler(null, dir);
+      });
   }
 };
 
@@ -72,6 +81,9 @@ let move = function(req, res, isFile, next) {
     req.app.get('api').nfs.moveFile(reqBody.srcPath, srcRootPath, reqBody.destPath, destRootPath,
       action, sessionInfo.hasSafeDriveAccess(), sessionInfo.appDirKey, responseHandler);
   } else {
+    if (action === false && reqBody.srcPath === '/') {
+      return next(new ResponseError(400, 'Cannot move root directory'));   
+    }
     log.debug('NFS - Invoking move directory request');
     req.app.get('api').nfs.moveDir(reqBody.srcPath, srcRootPath, reqBody.destPath, destRootPath,
       action, sessionInfo.hasSafeDriveAccess(), sessionInfo.appDirKey, responseHandler);
@@ -89,6 +101,9 @@ export var createDirectory = function(req, res, next) {
     return next(new ResponseError(400, util.format(MSG_CONSTANTS.FAILURE.FIELD_NOT_VALID, 'rootPath')));
   }
   let dirPath = req.params['0'];
+  if (!dirPath || dirPath === '/') {
+    return next(new ResponseError(400, 'Directory path specified is not valid'));
+  }
   reqBody.metadata = reqBody.metadata || '';
   reqBody.isPrivate = reqBody.isPrivate || false;
   if (typeof reqBody.metadata !== 'string') {
@@ -97,7 +112,6 @@ export var createDirectory = function(req, res, next) {
   if (typeof reqBody.isPrivate !== 'boolean') {
     return next(new ResponseError(400, util.format(MSG_CONSTANTS.FAILURE.FIELD_NOT_VALID, 'isPrivate')));
   }
-  let appDirKey = sessionInfo.appDirKey;
   log.debug('NFS - Invoking create directory request');
   let responseHandler = new ResponseHandler(req, res);
   req.app.get('api').nfs.createDirectory(dirPath, reqBody.isPrivate, false,
@@ -125,6 +139,9 @@ export var modifyDirectory = function(req, res, next) {
     return next(new ResponseError(400, util.format(MSG_CONSTANTS.FAILURE.FIELD_NOT_VALID, 'rootPath')));
   }
   let dirPath = req.params['0'];
+  if (!dirPath || dirPath === '/') {
+    return next(new ResponseError(400, 'Directory path specified is not valid'));
+  }
   reqBody.name = reqBody.name || '';
   reqBody.metadata = reqBody.metadata || '';
 
@@ -266,12 +283,12 @@ export var getFile = function(req, res, next) {
       "Content-Range": "bytes " + start + "-" + end + "/" + total,
       "Accept-Ranges": "bytes",
       "Content-Length": chunksize,
-      "Created-On": new Date(fileStats.createdTimeSec || fileStats.createdOn).toUTCString(),
-      "Last-Modified": new Date(fileStats.modifiedTimeSec || fileStats.modifiedOn).toUTCString(),
+      "Created-On": fileStats.createdOn,
+      "Last-Modified": fileStats.modifiedOn,
       "Content-Type": mime.lookup(filePath) || 'application/octet-stream'
     };
     if (fileStats.metadata) {
-      headers.metadata = fileStats.metadata;
+      headers.metadata = new Buffer(fileStats.metadata, 'base64').toString('base64');
     }
     res.writeHead(range ? 206 : 200, headers);
     if (chunksize === 0) {
@@ -303,9 +320,9 @@ export var getFileMetadata = function(req, res, next) {
       }
       res.writeHead(200, {
         "Accept-Ranges": "bytes",
-        "Created-On": new Date(fileStats.createdOn).toUTCString(),
-        "Last-Modified": new Date(fileStats.modifiedOn).toUTCString(),
-        "Metadata": fileStats.metadata,
+        "Created-On": fileStats.createdOn,
+        "Last-Modified": fileStats.modifiedOn,
+        "Metadata": new Buffer(fileStats.metadata, 'base64').toString('base64'),
         "Content-Type": mime.lookup(filePath) || 'application/octet-stream',
         "Content-Length": fileStats.size
       });
